move ProjectEndpoints to admin.api package
