+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/ProjectsEndpoints.scala
@@ -3,24 +3,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.knora.webapi.routing.admin
+package org.knora.webapi.slice.admin.api
 
+import org.knora.webapi.messages.admin.responder.projectsmessages._
+import org.knora.webapi.routing.BaseEndpoints
+import org.knora.webapi.routing.PathVariables.{projectIri, projectShortcode, projectShortname}
+import org.knora.webapi.slice.admin.api.model.{ProjectExportInfoResponse, ProjectImportResponse}
 import sttp.capabilities.pekko.PekkoStreams
 import sttp.model.StatusCode
 import sttp.tapir._
 import sttp.tapir.generic.auto._
 import sttp.tapir.json.spray.{jsonBody => sprayJsonBody}
 import sttp.tapir.json.zio.{jsonBody => zioJsonBody}
-import zio.Chunk
-import zio.ZLayer
-
-import org.knora.webapi.messages.admin.responder.projectsmessages._
-import org.knora.webapi.routing.BaseEndpoints
-import org.knora.webapi.routing.PathVariables.projectIri
-import org.knora.webapi.routing.PathVariables.projectShortcode
-import org.knora.webapi.routing.PathVariables.projectShortname
-import org.knora.webapi.slice.admin.api.model.ProjectExportInfoResponse
-import org.knora.webapi.slice.admin.api.model.ProjectImportResponse
+import zio.{Chunk, ZLayer}
 
 final case class ProjectsEndpoints(
   baseEndpoints: BaseEndpoints
